modify

+++ src/main/java/com/mybookmark/mybookmarkapi/domain/bookmark/BookmarkService.java
@@ -1,20 +1,15 @@
 package com.mybookmark.mybookmarkapi.domain.bookmark;
 
+import java.security.Principal;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 
-import javax.servlet.http.HttpServletRequest;
-
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.ResponseStatus;
 
-import com.mybookmark.mybookmarkapi.common.error.handler.ErrorResponse;
+import com.mybookmark.mybookmarkapi.domain.user.UserEntity;
+import com.mybookmark.mybookmarkapi.domain.user.UserRepository;
 import com.mybookmark.mybookmarkapi.domain.util.converter.DtoEntityMapper;
 
 @Service
@@ -23,8 +18,11 @@ public class BookmarkService {
 	@Autowired
 	private BookmarkRepository bookmarkRepository;
 	@Autowired
+	private UserRepository userReository;
+	@Autowired
 	private DtoEntityMapper dtoEntityMapper;
-
+	
+	
 	public BookmarkDto readBookmark(long bookmarkId) {
 		BookmarkEntity entity = bookmarkRepository.findByBookmarkId(bookmarkId);
 		if (entity != null) {
@@ -43,16 +41,21 @@ public class BookmarkService {
 		return dtos;
 	}
 
-	public void createBookmark(BookmarkDto dto) {
+	public void createBookmark(BookmarkDto dto, Principal principal) {
+		UserEntity user = userReository.findByLoginId(principal.getName());
 		BookmarkEntity entity = dtoEntityMapper.fromDtoToEntity(dto, BookmarkEntity.class);
+		entity.setUser(user);
 		bookmarkRepository.save(entity);
 	}
 
-	public boolean updateBookmark(long bookmarkId, BookmarkDto dto) {
+	public boolean updateBookmark(long bookmarkId, BookmarkDto dto, Principal principal) {
 		BookmarkEntity hasEntity = bookmarkRepository.findByBookmarkId(bookmarkId);
-		if (hasEntity != null) {
+		UserEntity registered = userReository.findByLoginId(principal.getName());
+		
+		if (hasEntity != null && registered.getLoginId().equals(hasEntity.getUser().getLoginId())) {
 			BookmarkEntity entity = dtoEntityMapper.fromDtoToEntity(dto, BookmarkEntity.class);
 			entity.setBookmarkId(bookmarkId);
+			entity.setUser(registered);
 			bookmarkRepository.save(entity);
 			return true;
 		} else {
@@ -60,8 +63,14 @@ public class BookmarkService {
 		}
 	}
 
-	public void deleteBookmark(long bookmarkId) {
-		bookmarkRepository.deleteById(bookmarkId);
+	public boolean deleteBookmark(long bookmarkId, Principal principal) {
+		BookmarkEntity bookmark = bookmarkRepository.findByBookmarkId(bookmarkId);
+		UserEntity registered = userReository.findByLoginId(principal.getName());
+		if (bookmark != null && registered.getLoginId().equals(bookmark.getUser().getLoginId())) {
+			bookmarkRepository.deleteById(bookmarkId);
+			return true;			
+		} else {
+			return false;
+		}
 	}
-
 }
